apfsck: make file descriptor into global variable
The file descriptor of the device is always the same, and only used in
the calls to mmap().  Make it into a global variable to avoid having to
pass it around.

With this change a few functions will have no parameters and return
void.  Use 'void' in the parameter list as well, to silence Sparse
compaints about 'non-ANSI function declarations'.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>