apfsck: call read_node() with oid instead of bno
Modify read_node() to be called with the object id of a node instead
of a block number.  Check that the oid in the metadata is correct
before returning, instead of doing it in the caller.  This simplifies
the code and, more importantly, extends that check to the root node.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>