apfsck: harden boundary checks for keys and values
The current checks performed by node_locate_key() and node_locate_data()
were copied from the kernel module.  Their purpose was not to detect
corruption, but to prevent out-of-bounds operations.  In the context of
the fsck tool, it makes sense to check that keys and values not only fit
in the block, but that they fit in their respective areas.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>