apfsck: verify that free node space is unused
For each node, compare the allocation bitmaps compiled from the free
space lists against the ones compiled from the table of contents.  That
is, free space records vs. used space records.

No free bytes can be in use, but unused bytes are not always listed as
free.  The reason is fragmentation: some free areas may be too tiny to
be part of a linked list.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>