apfsck: check object types
Read the type of an object in read_object(), and have the caller check
that it's correct.  Since all callers need to do this, stop allowing
read_object() to be called with obj == NULL.

For now, ignore the container superblock.  Most of that code will need
to be reworked when we start verifying all checkpoints instead of just
the latest.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>