apfsck: don't normalize key names for comparison
When two filenames have the same hash, compare them as character strings
to decide the order of their keys.  That is, ignore the case-folding and
unicode decomposition.

This has only been tested for ascii names, so some uncertainty remains.
Either way, the change is an improvement and should also be applied to
the module.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>