apfsck: compare an omap node xid with its child's
When a btree node is updated, the copy-on-write mechanism means that
the block number will change.  This is not important for virtual objects
such as catalog nodes, because their parent nodes only list object ids
that still need to be translated.

But for physical objects, like the nodes of an object map, this means
that any update of a node will force an update of its parent, so that
the block number in the nonleaf record can change.  Transaction ids must
therefore always be greater in a parent node.  Add a check for that.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>