apfsck: count links and children for each inode
Keep track of the number of dentry records linking to each inode, and of
the number of dentries inside each directory.  Check that they are sane,
and consistent with the inode metadata.  Since this can only be done
after the whole catalog has been parsed, let free_inode_table() take
care of it.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>