apfsck: verify that dentry type matches inode mode
The mode of an inode should match the file type of its dentries, except
for a bit shift.

Several dentries may be read before the inode, so save the type into the
higher bits of the i_mode field as soon as we know it.  Check that those
are the same for every dentry, and for the inode itself.  When the inode
is reached, complete the lower bits.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>