apfsck: keep a list of all extents in a volume
While checking the catalog tree, add an entry to a hash table for every
file extent record found.  For now its only purpose is to verify that
extents are consecutive, i.e., that each extent begins where the last
one ended.  More checks will be added in the future.

The hash table code is essentially the same for extents and inodes, so
I intend to refactor this soon.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>