apfsck: verify the name of an inode's single link
In contradiction with the official documentation, it seems that inodes
with a single link are also required to report their name in an xfield.
Adding a check for this is somewhat harder than may seem, because these
inodes don't always keep a list of sibling links.  Save the name of the
first dentry encountered for the inode, and use that if no other links
show up.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>