apfsck: add cli options to pick the checks to run
Often enough, there is no clear way to decide if something should be
reported as corruption.

Some filesystems created by Apple software have been found to violate
the specification in small ways (e.g., the volume file count).  This is
most likely not corruption, but I have no idea what can be done to
improve the check and I don't want to drop it.

Filesystems that have recently crashed will have some features that
should never be seen under normal use.  It may be useful to report when
they appear, to check if the module is unmounting cleanly.

Not all features of APFS will be supported in the near future, by either
the module or the fsck tool.  If any of them appear to show up after the
filesystem was used on linux, that's clearly something that should be
reported, even if it's technically not corruption.

To deal with this consistently, distinguish these three types of checks
and add command line options to let the user decide which ones to run.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>