apfsck: check consistency of inode flags
Check that the internal flags of an inode are all valid, and consistent
with each other.  Also report if any of them are unsupported, or if they
were set as the result of a crash.

Ignore the finder info flag for now, because it seems to be set in some
inodes that have no corresponding xfield.  The next patch will take care
of reporting this as weird.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>