apfsck: compare inode flags with extended fields
Some of the internal flags of an inode announce the presence of a given
type of xfield.  No xfield exists without its corresponding flag.  The
converse doesn't seem to hold for the finder info, so only report its
failures as weird.

Several checks are now performed on the xfield type bitmap, so add two
new helper functions, xbmap_test() and xbmap_set().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>