apfsck: support resource forks
I was under the impression that resource forks would be tricky to check,
so the fsck tool is reporting their presence as an unsupported feature.
But it seems that resource forks are actually just another xattr, under
the name "com.apple.ResourceFork".

Check that all inodes with the HAS_RSRC_FORK flag set have this xattr,
and vice versa.  Note that, unlike symlink targets, these xattrs are not
system-owned.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>