apfsck: move inode size information to dstream
Move the i_size and i_alloced_size fields from the inode structure to
the associated dstream.  This makes more sense because several different
inodes can share the same dstream, and xattrs can use them as well.

Check that every inode with the same private id reports the same size
for the dstream, and that no dstream is used by both an inode and a
xattr.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>