apfsck: plug leak of b-tree nodes
When parse_subtree() is done with a node, it should release it with
free_node() instead of just free().  Otherwise the associated free space
bitmaps will leak, and the physical block will remain mapped.

I have been aware of this issue for a while, but memory leaks are not
particularly urgent for a program that will only run for a few minutes
at a time.  There are still many structures I don't even attempt to
free.  Much of this will need to be improved when we start parsing old
checkpoints.

Since in-memory catalog keys could point to memory-mapped on-disk names,
make sure they are copied into a memory buffer before releasing their
node.  I'm not entirely pleased with this, but I can't think of a
cleaner solution without a big rewrite.

Fixes: a71899fe1353 ("apfsck: check order of keys in the container omap")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>