apfsck: don't create dstreams when checking inodes
Since it was first added, check_inode_stats() has been creating dstream
structures for every inode it checks.  This was not a problem when we
were only checking the size, because the newly created dstream would be
empty anyway.

But when checks were added for the number of references, I started to
notice a large number of dstreams that seemed to have none.  I thought
this was a sign of some sort of garbage collection and put messy checks
in place to deal with it.

Stop calling get_dstream() from check_inode_stats(), and instead keep a
pointer to the dstream in the inode structure.  Report dstreams with no
references as corruption.

Fixes: 647278ebfa25 ("apfsck: check dstream id records")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>