apfsck: compare object xid with first volume xid
Check that the transaction id of all objects inside a volume comes after
the first one, i.e., the one reported by apfs_formatted_by.

Note that it isn't true in general that the last transaction matches the
latest entry in apfs_modified_by, because keeping those updated is not
mandatory.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>