apfsck: check if the container fits in the device
Check that the block count reported by the container is not bigger than
the block count of the device being checked.

The headers that must be included for fstat() and ioctl() contain new
definitions for a few macros I was already using.  To avoid conflicts,
get rid of my versions.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>