apfsck: check if the container reports a keybag
Containers with no encryption may still have non-zero values in the
nx_keylocker field (which reports the location of the keybag).  There is
no actual keybag in the given location, so I assume they are ignored.

I don't understand why this happens and I don't support encryption, so
just report it as weird.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>