apfsck: check container's array of ephemeral info
The container superblock has an array of fields supposedly used in the
management of ephemeral data.  Only the first field is documented; the
others may not be in use.  Add the proper checks.

The new check_ephemeral_information() function expects sb->s_block_count
to be set before the call.  Use an assertion to remind the programmer of
this; to make sure it's effective, initialize the in-memory superblock
with zeroes when allocated.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>