apfsck: add extentref checks to parse_subtree()
Check that the nodes of the extent reference tree have variable sized
keys, like the nodes of the catalog.  This is strange, but it seems to
be true.

Also, since these nodes are physical objects (like the omap nodes),
check that the transaction ids of children are always below the xid of
the parent.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>