apfsck: add blocksize_bits to in-memory superblock
Keep a blocksize_bits field in the in-memory superblock structure, like
the linux vfs does.  The point of this field is to remember how many
bits need to be shifted to translate a disk address to a block number.

For the calculation, copy the code from the kernel and add consistency
checks for the block size.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>