apfsck: check reference count of physical extents
Check that the refcounts provided by the extent reference tree are
correct, at least to the best of my understanding.  This patch was
tricky to get right, because the documentation is lacking in important
details.  For one, it doesn't mention that you should only count one
reference when a physical extent is shared by several xattrs of the same
inode.

In general, the behaviour of clones is not sufficiently tested, because
I have no access to appropriate filesystem images.  This is very likely
to cause trouble.

The implementation of the check is also far less clean than I would have
hoped.  Keep the physical extents in a hash list like usual, but also
keep them in a linked list for each data stream.  When the data stream
is freed, increase the reference count of each physical extent only
once.  Pay attention to the owner of the dstream, to avoid counting
xattrs incorrectly.

We often need to deal with file extents that refer to an address in the
middle of a physical extent.  The easiest way is to bypass the hash
table and search the on-disk b-tree directly for the closest match;
adjust the query code for this purpose.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>