apfsck: check that owner ids are in valid range
If the owner of a physical extent is removed, I don't see how the owning
object id could be updated to refer to another of the users.  My guess
is that it will be left as it is, so checking these ids for correctness
won't be possible in general.

Just confirm that they are in the valid range.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>