apfsck: fix data boundary check in node_is_valid()
When a node has no records, the beginning of the data area will match
the end of the block or the beginning of the footer.  Currently, such a
thing would be reported as corruption, but it's not a problem because
empty nodes are not being considered at all.  This is about to change
with the addition of the snapshot metadata tree, so get it fixed.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>