apfsck: unmap on-disk structs after read_object()
When parse_omap_btree(), parse_reaper() and check_efi_information()
return, the pointer to their raw on-disk structures is discarded and the
memory mapping leaks.  A handful of leaks such as these won't really
matter, especially since the program only runs for a short time, but do
the right thing and call munmap() anyway.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>