apfsck: move read_ephemeral_object() to object.c
Move the definition of the read_ephemeral_object() function to object.c
and declare it as extern, so that it can later be called from the space
manager code.  Also move free_cpoint_map(), free_cpoint_map_table() and
get_cpoint_map(), since it makes sense to keep them together.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>