apfsck: check chunk and cib counts for main device
Check that the count of chunks and chunk-info blocks reported by the
main device is consistent with the size of the container.  Preserve the
values in the superblock structure because they will need some further
checks later.

Since chunk-info address blocks are not supported, the cab count
reported by the main device should always be zero.  Add a check for
that as well.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>