apfsck: define an in-memory spaceman structure
The space manager already has several fields in the in-memory superblock
structure, and it will need many more in the future.  Wrap them in a new
spaceman structure.  Rename existing variables for the raw spaceman from
'sm' to 'raw', so that the first can be used for the in-memory struct
instead.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>