apfsck: read disk allocation bitmap into memory
Read the bitmaps for each individual chunk and stitch them together in
memory as a single large bitmap for the whole container.  This will
later be checked against the bitmap of actual used blocks, but for now
only verify the free block count as a sanity check for the patch.

Take advantage of the __builtin_popcountll() builtin, but don't bother
dealing with the different cpus unless we actually see a performance
improvement.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>