apfsck: check block count and xids of free queues
Check that the total free block count reported by a free queue's records
and the first xid among them both match the values provided in the free
queue structure itself.

We need a place to store those stats as we parse the free queue b-tree,
so define a new free_queue structure as a subclass of the btree struct.
Pass a pointer to the b-tree to parse_free_queue_record(), which can
then simply be cast to free_queue.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>