apfsck: begin to assemble the allocation bitmap
Add a new s_bitmap field to the in-memory container superblock, pointing
to a large buffer that will, after a few more patches, hold the real
allocation bitmap for the filesystem.  Begin by marking block zero (the
superblock backup) as used.

After the reported allocation bitmap is read from the space manager, run
the new compare_container_bitmaps() to verify that it's not inconsistent
with s_bitmap.  For now we can only check that used blocks are marked as
such in the reported bitmap; the converse will be added in the near
future, after we identify every block in use.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>