apfsck: mark the free queue record blocks as used
The free queue b-tree is not very well documented, but it seems to be a
list of extents that recently became unused.  The point is that they are
not immediately marked as free in the allocation bitmap, so their
checkpoints are preserved for a while.

Continue to assemble the actual allocation bitmap by marking these
blocks as used.  Define container_bmap_mark_as_used() as a helper, to be
reused in the following patches.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>