apfsck: issue endianness warnings from sparse
When the definitions of little-endian types copied from the kernel were
replaced with those from <linux/types.h>, I failed to realize that I
also had to define the __CHECK_ENDIAN__ macro.  As a result, sparse has
given no endianness warnings for quite a while.

Fix this, along with all issues spotted when running the tool.

Fixes: 24d928bf6940 ("apfsck: check if the container fits in the device")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>