apfsck: complete compare_container_bitmaps()
The real allocation bitmap should be fully assembled now, so add the
converse check to compare_container_bitmaps().  That is, check that no
free block is marked as used in the on-disk bitmap.

One my testing images has a single used block that remains unaccounted
for.  This image also claims to have one extra file in the catalog tree,
so I'm thinking both issues may somehow be related.  For now just report
it as weird.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>