apfsck: check the bitmap list of the internal pool
The space manager's internal pool keeps a list of old bitmaps along with
the current one.  Verify that the fields describing the list are
consistent with each other, and that the current bitmap is not in the
free range.  Move all these checks to the new parse_ip_bitmap_list() to
keep check_internal_pool() simple.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>