apfsck: deal with free_next_offset
The sm_ip_bm_free_next_offset field of the space manager is undocumented
and I have failed to make sense of it.  So far it always seems to point
to the same value, so add a check for it, and a TODO comment so that I
don't forget about this problem.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>