apfsck: handle 'versioned' flag in space manager
It appears that the last three fields of the on-disk spaceman structure
are only available when the 'versioned' flag is set.  This settles the
confusion caused by zone boundaries sometimes being set even if the id
was not: as it turns out, those fields were not part of zone definitions
at all, so don't check them.

For reasons I ignore, one of these optional fields reports the size of
the spaceman itself.  Check that the value is correct.

While we are at it, check that all of the spaceman offsets resolved by
spaceman_val_from_off() come after the end of the structure itself.
This requires storing the actual size of the spaceman into the in-memory
structure before any call to the function.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>