apfsck: move raw struct definitions to common dir
Create a new common 'include' directory outside of 'apfsck', and move
all definitions for on-disk structures in there, as well as the always
useful types.h.  An 'apfs' subdirectory is used to avoid potential
collisions with the system headers.

The intention is to also be able to include these files from the soon to
come mkfs.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>