apfsck: report more warnings on build
I'm always passing the -Wall flag to gcc manually; hardcode into the
Makefile instead.  If Sparse is available, run it on every build, with
the same flags as gcc to make it aware of the new header location.

Sparse warnings would get lost among the current output of make, so echo
shorter messages instead, and indent them.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>