apfsck: move unicode routines to common directory
Create a new 'lib' directory outside of 'apfsck', and move all the
unicode routines in there.  Also move their headers to include/apfs.
The intention is to be able to link this file to the mkfs as well.

Write a Makefile for the new library directory, and call it from the
original Makefile for the fsck.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>