apfsck: move crc32c routines to common directory
Move the crc32c code to the lib directory, and its headers to include.
The intention is to be able to link this file to the mkfs as well.

Tweak the fsck Makefile one more time, to prevent the recursive call to
make from happening once for each library file.  From now on, always
relink apfsck, even if the libraries weren't updated.  This is not ideal
but it's good enough for now.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>