mkapfs: set object headers of container superblock
It's too soon to know the actual checksum for the container superblock,
since too many fields are not yet set.  Start setting the object headers
anyway, so that the fsck will take longer to fail and allow some more
testing.

Define set_object_header() for this purpose; it will later be reused
for all other objects.  Make the filesystem parameters into a global
variable to avoid having to pass them around.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>