mkapfs: make the root node of the container's omap
Define make_main_omap_root() to make a root node for the container's
object map, including a single record for mapping the first volume's oid
to its block number.

Some of this code will probably be moved to separate functions in the
future, so that it can be reused for other trees.  Something like that
was already done with set_omap_info().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>