mkapfs: set feature flags on volume superblock
Set the flags for the volume features in the superblock.  By default the
volume is case-sensitive, but add and document a command line flag to
override this.  It's the first cli option to refer to a volume instead
of the container, so make it clear in the man page that we only support
single-volume containers, for now.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>