mkapfs: write the UUID of the volume
Handle the UUID of the volume like the container's:  allow the user to
choose it via cli, otherwise use a random value.  Reuse set_uuid() and
define a new get_random_uuid() to call for both in the default case.

While moving code into this new function, fix a couple of bugs that
could be triggered if reads from the /proc filesystem failed, which I'm
guessing is unlikely.  Also remember to close the file descriptor.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>