mkapfs: write volume creation information
Record the time of the volume creation in its superblock, with the best
precision available.  The timestamp must be provided in nanoseconds, so
make use of the NSEC_PER_SEC macro from the Linux kernel.  Put the code
in a get_timestamp() inline function so that other filesystem records
can take advantage of it in the future.

Record the identity of this mkfs tool as well, and the transaction id.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>