mkapfs: make an object map for the volume
Extend make_main_omap_root() (now 'make_omap_root') and
make_omap_btree() to also make an object map for the volume, with a
single record mapping the root node of the catalog tree.

I'm not very pleased with the quality of the code here, but most of the
hardcoded stuff will go away eventually, when we add support for
multiple volumes.  This is not likely to happen before the module gets
write support.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>