mkapfs: make the catalog root node
Make the root of the catalog tree and place four records in it: a dentry
and an inode for each of the root and private directories.  The code is
long and messy, and continues the awkward use of hardcoded constants.
I'd rather not put much effort in improving it, since I expect most of
it to be replaced after write support is implemented in the module.

The fsck was originally reporting a bad directory count on images
generated with this patch.  I had forgotten that special inodes are
ignored by that field, so the count should actually be zero, not two.
Fix this as well.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>