mkapfs: make the free queues for the spaceman
Make and map two empty free queues: one for the internal pool, and one
for the main device.  Use make_empty_btree_root() to make the root node.

It was always the plan to use this function for the free queues as well,
but some issues were still discovered.  The free queue nodes are not
physical objects, so the object id needs to be passed to the function.
The value lengths in the info footer were not correct either.  Get this
fixed.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>