mkapfs: make the allocation bitmap
Use the hardcoded block numbers to make an allocation bitmap for the
first chunk, which is the only one used by the mkfs.  Copy and simplify
container_bmap_mark_as_used() from the fsck, to be used as a helper in
the process.

Also, define count_used_blocks() in a similar style, and use it to set
the free block counts of both chunk and device to the correct values.

With this patch the mkfs is finally capable of producing a consistent
filesystem.  We are ready to move back to the kernel module and start
working on write support.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>