apfsprogs: handle meta_crypto fields in volume sb
I have no intention of supporting encryption anytime soon, so the
related structures have mostly been ignored so far.  The meta_crypto
structure in the volume superblock needs some attention though, because
it's always present and not set to zero.

Follow the documentation to define make_meta_crypto() in the mkfs, and
check_meta_crypto() in the fsck.  Use WMCS_PROTECTION_CLASS_F, which
seems to mean 'no encryption'.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>