apfsprogs: handle inode's default protection class
In the fsck, check that the default protection class of an inode takes
one of the two values that have been encountered in unencrypted volumes:
APFS_PROTECTION_CLASS_DIR_NONE, meaning the protection is inherited from
the parent directory, or APFS_PROTECTION_CLASS_D, meaning no protection.
I'm guessing APFS_PROTECTION_CLASS_F may also be allowed, but I can't
confirm this so ignore it for now.

In the mkfs, just set the class of every inode to
APFS_PROTECTION_CLASS_DIR_NONE.  The value is zero, so this changes
nothing.

The previous patch defined the protection class macros directly in the
super.c files, right before they were used; I was hoping to keep this
simple because encryption is not supported.  Now that several classes
are in use, it's best to just place all the macros in the common raw.h
header.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>