mkapfs: wipe old checkpoint superblocks
If the disk was ever formatted as APFS before, it is possible that valid
checkpoint superblocks remain in the descriptor area.  Their transaction
ids are likely to be higher than MKFS_XID, so one of them will always be
mounted instead of the right one.

Avoid this by wiping the whole checkpoint descriptor area before working
on it.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>