apfsck: use latest superblock instead of backup
parse_filesystem() only verifies the checkpoints reported as valid by
the superblock in block zero.  If it's not up-to-date, the valid range
will be obsolete as well, and the call to main_super_compare() won't
notice the crash.

Check the range provided by the latest superblock instead.  Don't let
main_super_compare() run if the backup superblock is old.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>