apfsck: rework the hash tables
Clean up our hash table implementation.  Get rid of that nonsense with
unions and just cast the headers to the whole entry when needed.  Embed
the actual header structure in the entries instead of reproducing it,
and use a macro to allow access to the object ids in the old fashion.

Get rid of the redundant TABLE_BUCKETS macros, since all hash tables use
the same value.

I haven't given any thought to the potential strict aliasing issues
here, so just add the -fno-strict-aliasing flag to all Makefiles.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>