apfsprogs: correct the dentry hash mask
The official documentation defined APFS_DREC_HASH_MASK incorrectly.
This bug got copied into both apfsprogs and the kernel module, which
fixed it long ago with the commit d65ef2c60d16 ("apfs: correct the
dentry hash mask").  So far the macro is unused here, but fix it so
that the same header can be used in both kernel and userland.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>