apfsprogs: remove an in-memory struct from raw.h
The definition of the in-memory extref_record structure has been placed
in raw.h by accident.  Move it to extents.h in the apfsck directory.
It's a bit odd that we need extref_record when we already have an extent
struct, so add a little comment on its purpose.

The apfs_raw.h file in the kernel module will also have to be updated.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>