apfsck: allow different flags in the backup sb
I'm looking at a fresh filesystem image created with an iphone vm, and
the nx_flags field in the superblock backup doesn't match the one from
the latest checkpoint. The extra flag in the copy is APFS_NX_RESERVED_2
which, of course, is undocumented.

I have no idea what this is about, but don't report it as corruption.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>