apfsck: fix calculation of free blocks in ip ring
Regarding the ring buffer of bitmaps for the internal pool, I mistakenly
thought that the tail of the free area was free itself, when it's not.
This means that regular containers have two non-free bitmaps in the ring
buffer, not just one.

Fix the calculation of the length, which will be important when studying
the ip_bm_free_next bitmap.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>