apfsck: check that the ip has 16 bitmaps
Despite having a field to report its length, the internal pool appears
to always have 16 bitmap blocks. My guess is that bigger containers may
need two blocks per bitmap, and so they may have 32 blocks in total. But
such containers are not supported so far and I have never seen one, so
report this as corruption.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>