apfsck: add checks for the ip_bm_free_next field
This field is undocumented, but it appears to distinguish between used
and free bitmaps in the ip ring buffer. Of course this can also be done
by checking the head and tail of the free area, so I fail to see the
point. Anyway, add the proper checks; the module will also need a patch.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>