mkapfs: preallocate toc for fixed kv size nodes
As explained in commit 7f570159e351 ("apfsck: check preallocated toc for
fixed kv nodes"), when a node has keys and values of a fixed size,
implementations are expected to preallocate a table of contents that
covers the whole node. This was already fixed in the module, but the
mkfs has been neglected for a while; deal with it now.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>