apfsprogs: deal with maximum size of free queues
The Apple implementation refuses to mount any container that doesn't
have a specific tree node limit set in the free queues. It took me a
while, but I think I figured out the formulas they use. Implement them
in a new common library file, and use it both for the fsck to check the
value and for the mkfs to set it correctly in new filesystems.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>