mkapfs: fix endianness issue in ip free_next field
I recently introduced an endianness issue in the mkfs. I'm not really
serious about supporting big-endian architectures in userland; some work
would be required for that, fixing several bitmaps and offset fields.
But I did set the Makefile to run sparse on build, and it's complaining
about this one issue, so just get it fixed.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>