mkapfs: fix check of label length
The length of the label provided by the user is checked early on, but
Luflosi has noticed that I forgot to exit if the check fails:

  https://github.com/linux-apfs/apfsprogs/issues/5

This could lead to a segfault later, when the label is copied into the
volume superblock. There are other similar problems in this program that
I have chosen to ignore so far, but this one is clearly a mistake and is
easy to fix, so do it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>