apfsck: allow rdev to be zero for device files
For some silly reason I thought that device files weren't allowed to
have a null device id; that check causes the fsck to report corruption
after running generic/013 of xfstests. Remove it, and replace it with a
check that all device files (even the null ones) have an rdev xfield. I
have confirmed this is true for the official implementation.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>