mkapfs: support larger containers
The size of the allocation bitmap is hardcoded as only one block. This
is not enough for containers with a large internal pool, so the bitmap
write overflows with unpredictable results. I've been aware of this
problem for a while and didn't think it was urgent, but I decided to get
it fixed after I got a bug report about it:

  https://github.com/linux-apfs/linux-apfs-rw/issues/17

Just calculate the number of chunks that will be allocated, and use as
large a bitmap as needed.

Note that this is not enough to truly support containers over 1 Tb in
size. Those might work fine with my module alone, but the official
implementation won't like them. The problem is that we don't yet support
multiblock allocation bitmaps for the internal pool.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>