apfsck: move volume checking code to new functions
Snapshots have their own internal volume superblocks, which are similar
to the current one but also have some important differences, for example
being physical objects instead of virtual. To avoid having to copy-paste
all the checks, move them into a couple of new functions that can later
be reused for the snapshots.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>