apfsck: parse the omap snapshots tree
While trying to track down the source of two extra blocks in a volume's
apfs_fs_alloc_count, I noticed that there is an extra snapshot tree
related to the object map. I'm not sure why it's there, but it seems
simple enough so add checks for it.

This only accounts for one of the two extra blocks though, so I'll have
to keep looking.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>