apfsck: fix physical extent checks for snapshots
Checks for the extent reference trees run into a few problems when
snapshots are present. First of all, a volume's apfs_fs_alloc_count
only includes its own extents, not the ones of its snapshots. This
surprised me a little, because the snapshot is otherwise considered a
part of the volume. To deal with it, don't count the extent blocks when
the physical record is read, but instead do it when it's attached to a
dstream.

More problems come from the presence of a second kind of physical extent
record: the update record. These records are created when the current
transaction needs to modifiy the reference count of a snapshot extent:
the original record in the snapshot's extentref tree is untouched, while
an update record gets added to the current extentref tree.

Don't count update records for apfs_fs_alloc_count, their blocks belong
to the snapshot. Don't mark those blocks as used in the allocation
bitmap either, the snapshot will handle them. Most importantly, fix the
calculation of the reference count to include the updates.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>